Modification

+++ 3.uloha.py
@@ -1,4 +1,4 @@
-x=float(input("Input number: "))
+x=int(input("Input number: "))
 q=1/x
 
 print(q)
